detect/http_cookie: move tests into tests/
